Moved the @Wrapped annotation into the transpiler tests package
