+++ transpiler/src/test/java/source/extension/Wrapped.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package jsweet.lang;
+package source.extension;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
